src/requirements.txt
alembic==1.8.1
attrs==19.3.0
Automat==0.8.0
autopep8==1.7.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
click==8.1.3
cloud-init==20.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
Flask==2.2.2
Flask-Dance==6.1.1
Flask-Login==0.4.1
Flask-Migrate==3.1.0
Flask-Seeder==1.2.0
Flask-SQLAlchemy==3.0.0
greenlet==1.1.3
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==5.0.0
importlib-resources==5.9.0
incremental==16.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
Mako==1.2.3
MarkupSafe==2.1.1
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.0.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.9.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.4.2
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0
python-debian===0.1.36ubuntu1
python-dotenv==0.21.0
PyYAML==5.3.1
requests==2.21.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
SQLAlchemy==1.4.41
ssh-import-id==5.10
systemd-python==234
toml==0.10.2
Twisted==18.9.0
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.24.3
URLObject==2.4.3
wadllib==1.3.3
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1
zope.interface==4.7.1
